Revert "Merge pull request #4475 from bbc/t2347-taken-date-ui"
This reverts commit 1322691719fe180a6004398880bf265007c3c3cf, reversing
changes made to d7e37a68f8ab69940157debba930bf8c9e7926db.
