+++ kahuna/public/js/components/gr-panels/gr-panels.js
@@ -69,13 +69,7 @@ panels.directive('grPanel', ['$timeout', '$window', 'inject$', 'subscribe$',
 
             // Then hide the panel
             subscribe$(scope, scrollWhileVisAndUnlocked$, () => {
-                scope.$apply(() => {
-                  panel.setHidden(true);
-                  window.dispatchEvent(new CustomEvent("panelHide", {
-                    detail: {panel: "scroll"},
-                    bubbles: true
-                  }));
-                });
+                scope.$apply(() => panel.setHidden(true));
             });
         }
     };
