+++ kahuna/public/js/components/gr-search-wrapper/gr-search-wrapper.html
@@ -26,4 +26,6 @@
 
     <my-uploads class="my-uploads" props="searchQuery.myUploadsProps"></my-uploads>
 
+    <sort-control class="sort-control" props="searchQuery.sortProps"></sort-control>
+
 </div>
