--- kahuna/public/js/components/gr-tab-swap/gr-tab-swap.css
@@ -1,43 +1,0 @@
-.gr-tab-wrapper {
-  padding: 4px;
-}
-
-.gr-tab-panel-margin {
-  margin-left: 250px;
-}
-
-.gr-tab-container {
-  font-family: "Open Sans", sans-serif;
-  display: flex;
-  width: fit-content;
-}
-
-.gr-tab {
-  cursor: pointer;
-  font-size: 14px;
-  transition: background-color 0.3s ease;
-  padding: 8px 16px;
-  background-color: #444;
-  color: #ccc;
-  border: none;
-  outline: none;
-  position: relative;
-}
-
-.gr-tab:hover {
-  color: white;
-}
-
-.gr-tab.active {
-  background-color: #666;
-}
-
-.gr-tab.active::before {
-  content: "";
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 3px;
-  background-color: #00bfff;
-}
