+++ kahuna/public/js/search/results.html
@@ -21,7 +21,7 @@
             </gr-panel-button-small>
 
             <div class="results-toolbar-item results-toolbar-item--static
-                    image-results-count">
+                        image-results-count">
                 {{ctrl.totalResults | toLocaleString}} matches
                 <button ng-if="ctrl.newImagesCount > 0"
                         ng-click="ctrl.revealNewImages()"
@@ -73,11 +73,11 @@
                 </div>
 
                 <div class="results-toolbar-item results-toolbar-item--right results-toolbar-item--static results-toolbar-item__progress"
-                     ng-style="ctrl.buildBatchProgressGradient()" ng-if="ctrl.batchOperations.length > 0">
+                    ng-style="ctrl.buildBatchProgressGradient()" ng-if="ctrl.batchOperations.length > 0">
 
-                <span ng-repeat="entry in ctrl.batchOperations">
-                    Updating {{entry.key}}: {{entry.completed}}/{{entry.total}}
-                </span>
+                    <span ng-repeat="entry in ctrl.batchOperations">
+                        Updating {{entry.key}}: {{entry.completed}}/{{entry.total}}
+                    </span>
                 </div>
 
                 <a class="results-toolbar-item results-toolbar-item--right"
@@ -98,8 +98,8 @@
                 </a>
 
                 <div class="results-toolbar-item results-toolbar-item--right"
-                     tabindex="0"
-                     ng-if="ctrl.showSendToPhotoSales() && ctrl.showPaid">
+                    tabindex="0"
+                    ng-if="ctrl.showSendToPhotoSales() && ctrl.showPaid">
                     <a id="send-to"
                        style="display: contents;"
                        ng-class="{'batch-archive__button--disabled': ctrl.selectionCount >= 45}"
@@ -115,12 +115,12 @@
 
 
                 <gr-batch-export-original-images class="results-toolbar-item results-toolbar-item--right"
-                                                 images="ctrl.selectedImages"
-                                                 ng-if="ctrl.selectionCount > 0"></gr-batch-export-original-images>
+                  images="ctrl.selectedImages"
+                  ng-if="ctrl.selectionCount > 0"></gr-batch-export-original-images>
 
                 <gr-downloader class="results-toolbar-item results-toolbar-item--right"
-                               images="ctrl.selectedImages"
-                               ng-if="ctrl.selectionCount > 0"></gr-downloader>
+                    images="ctrl.selectedImages"
+                    ng-if="ctrl.selectionCount > 0"></gr-downloader>
 
                 <gr-delete-image class="results-toolbar-item results-toolbar-item--right"
                                  images="ctrl.selectedImages"
@@ -128,8 +128,8 @@
                 </gr-delete-image>
 
                 <gr-un-delete-image class="results-toolbar-item results-toolbar-item--right"
-                                    images="ctrl.selectedImages"
-                                    ng-if="ctrl.selectionCount > 0 && ctrl.isDeleted">
+                                 images="ctrl.selectedImages"
+                                 ng-if="ctrl.selectionCount > 0 && ctrl.isDeleted">
                 </gr-un-delete-image>
 
                 <gr-archiver class="results-toolbar-item results-toolbar-item--right"
@@ -146,13 +146,7 @@
                 </gr-panel-button>
             </div>
         </div>
-
-        <div ng-if="ctrl.usePermissionsFilter()">
-            <extended-sort-control class="sort-control" props="ctrl.extendedSortProps"></extended-sort-control>
-        </div>
-
-        <ul ng-class="{'results': !ctrl.usePermissionsFilter(),
-                       'results--alternate': ctrl.usePermissionsFilter()}"
+        <ul class="results"
             gu:lazy-table="ctrl.imagesAll"
             gu:lazy-table-cell-min-width="280"
             gu:lazy-table-cell-height="303"
@@ -212,7 +206,7 @@
             </li>
         </ul>
 
-        <div ng-if="ctrl.totalResults > ctrl.maxResults && !ctrl.previewView"
+        <div ng-if="ctrl.totalResults > ctrl.maxResults"
              class="image-results-more">
             <div class="image-results-more__heading">Too many results to display</div>
             <div class="image-results-more__instructions">Please refine your search to limit the number of results</div>
