+++ kahuna/public/js/search/syntax/syntax.html
@@ -167,13 +167,5 @@
                 Search by person (email) or ftp source (folder).
             </dd>
         </dl>
-        <dl class="advanced-search-example">
-            <dt class="advanced-search-example__field">
-                <gr-chip-example gr-filter-field="has" gr-example-search="dateTaken"></gr-chip-example>
-            </dt>
-            <dd class="advanced-search-example__explanation">
-                Returns images that have a 'Taken date' value set.
-            </dd>
-        </dl>
     </div>
 </div>
