+++ kahuna/public/stylesheets/main.css
@@ -1241,13 +1241,6 @@ textarea.ng-invalid {
     position: relative;
 }
 
-.results--alternate {
-    display: flex;
-    flex-wrap: wrap;
-    top: 96px;
-    position: relative;
-}
-
 .results-controls {
     background: white;
     padding: 10px;
@@ -1287,7 +1280,7 @@ textarea.ng-invalid {
     display: none;
 
     /* above thumbnail and overlay */
-    z-index: 1;
+    z-index: 2;
 }
 
 .result__select--no-pointer-events {
@@ -1551,7 +1544,7 @@ textarea.ng-invalid {
 .image-no-results {
     font-size: 3rem;
     text-align: center;
-    margin-top: 10rem;
+    margin-top: 4rem;
 }
 
 .image-loading-results {
@@ -3087,7 +3080,7 @@ FIXME: what to do with touch devices
   text-align: left;
 
   /* above thumbnail and overlay */
-  z-index: 1;
+  z-index: 2;
 }
 .result__select__overlay__text.alert {
   background-color: red;
